Rebased to v2.5.1 (#156)
